+++ src/Krystal/Session/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.3
 ---
 
+ * Added `getFlashed()` to return flashed session keys (i.e the ones that will be removed after retrieval)
  * Added `hasMany()` to check several keys for existence at once
  * Added `setMany()` to set a collection at once
  * Since now `set()` returns a chain instead of void